Refactor permission, user and group
