+++ packages/serverless/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zkdb",
+  "name": "zkdb-serverless",
   "version": "0.1.1",
   "description": "zkDatabase for the future of Web3",
   "author": "chiro@orochi.network",
@@ -9,6 +9,7 @@
     "database",
     "zk-database",
     "zkdb",
+    "serverless",
     "zkdatabase"
   ],
   "repository": {
