+++ packages/serverless/src/model/abstract/database-engine.ts
@@ -1,21 +1,9 @@
 import { MongoClient, ObjectId, ServerApiVersion, Document } from 'mongodb';
 import logger from '../../helper/logger';
-
-export const ZKDATABASE_INDEX_RECORD = '_zkindex';
-export const ZKDATABASE_INDEX_COLLECTION = '_zkdatabase_index';
-// Group
-export const ZKDATABASE_GROUP_COLLECTION = '_zkdatabase_group';
-export const ZKDATABASE_GROUP_PERMISSION_COLLECTION =
-  '_zkdatabase_group_permission';
-// User -> Group mapping
-export const ZKDATABASE_USER_GROUP_COLLECTION = '_zkdatabase_user_group';
-// User permission
-export const ZKDATABASE_USER_PERMISSION_COLLECTION =
-  '_zkdatabase_USER_permission';
-// Settings
-export const ZKDATABASE_SETTING_COLLECTION = '_zkdatabase_setting';
-// System management
-export const ZKDATABASE_MANAGEMENT_DB = '_zkdatabase_management';
+import {
+  ZKDATABASE_INDEX_COLLECTION,
+  ZKDATABASE_INDEX_RECORD,
+} from '../../common/const';
 
 type ZKDatabaseIndex = {
   [ZKDATABASE_INDEX_RECORD]: number;
